+++ lib/constants.d.ts
@@ -1,4 +1,10 @@
 /**
+ * The default license filename. Used when `config.licenseLink` is set as
+ * `true`.
+ */
+export const DEFAULT_LICENSE_FILENAME: 'LICENSE';
+
+/**
  * The placeholder text for sections that have no defined body.
  */
 export const PLACEHOLDER: '_To be documented._';