+++ lib/linkify-license.d.ts
@@ -1,4 +1,4 @@
 /**
  * Takes the body of the license section and injects a link to the license file.
  */
-export default function linkifyLicense(licenseBody: string, link: sting): string;
+export default function linkifyLicense(body: string, link: sting): string;